Refactor SharedImageRepresentation to use size_t for plane index
This value is used to index into an array and so should be a size_t
instead of an int.

This fixes -Wsign-conversion warnings when this code is included by a
file compiled with //build/config/compiler:prevent_unsafe_narrowing.

Bug: 493652457
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7792039
Auto-Submit: Reilly Grant <reillyg@chromium.org>
Commit-Queue: Geoff Lang <geofflang@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1621874}

+++ gpu/command_buffer/service/shared_image/compound_image_backing.cc
@@ -160,7 +160,7 @@ class WrappedGLTextureCompoundImageRepresentation
     compound_backing()->NotifyEndAccess(wrapped_->backing(), access_mode_);
   }
 
-  gpu::TextureBase* GetTextureBase(int plane_index) final {
+  gpu::TextureBase* GetTextureBase(size_t plane_index) final {
     return wrapped_->GetTextureBase(plane_index);
   }
 
@@ -168,7 +168,7 @@ class WrappedGLTextureCompoundImageRepresentation
     return wrapped_->SupportsMultipleConcurrentReadAccess();
   }
 
-  gles2::Texture* GetTexture(int plane_index) final {
+  gles2::Texture* GetTexture(size_t plane_index) final {
     return wrapped_->GetTexture(plane_index);
   }
 
@@ -208,7 +208,7 @@ class WrappedGLTexturePassthroughCompoundImageRepresentation
     compound_backing()->NotifyEndAccess(wrapped_->backing(), access_mode_);
   }
 
-  gpu::TextureBase* GetTextureBase(int plane_index) final {
+  gpu::TextureBase* GetTextureBase(size_t plane_index) final {
     return wrapped_->GetTextureBase(plane_index);
   }
 
@@ -217,7 +217,7 @@ class WrappedGLTexturePassthroughCompoundImageRepresentation
   }
 
   const scoped_refptr<gles2::TexturePassthrough>& GetTexturePassthrough(
-      int plane_index) final {
+      size_t plane_index) final {
     return wrapped_->GetTexturePassthrough(plane_index);
   }
