+++ gpu/command_buffer/service/shared_image/gl_ozone_image_representation.cc
@@ -47,7 +47,7 @@ GLTexturePassthroughOzoneImageRepresentation::
 
 const scoped_refptr<gles2::TexturePassthrough>&
 GLTexturePassthroughOzoneImageRepresentation::GetTexturePassthrough(
-    int plane_index) {
+    size_t plane_index) {
   return textures_holder_->texture(plane_index);
 }
