+++ gpu/command_buffer/service/shared_image/shared_image_representation.cc
@@ -166,7 +166,7 @@ bool GLTextureImageRepresentationBase::SupportsMultipleConcurrentReadAccess() {
 // GLTextureImageRepresentation
 
 gpu::TextureBase* GLTextureImageRepresentation::GetTextureBase(
-    int plane_index) {
+    size_t plane_index) {
   return GetTexture(plane_index);
 }
 
@@ -199,7 +199,7 @@ void GLTextureImageRepresentation::UpdateClearedStateOnBeginAccess() {
 // GLTexturePassthroughImageRepresentation
 
 gpu::TextureBase* GLTexturePassthroughImageRepresentation::GetTextureBase(
-    int plane_index) {
+    size_t plane_index) {
   return GetTexturePassthrough(plane_index).get();
 }
 
@@ -516,7 +516,7 @@ SkiaGaneshImageRepresentation::ScopedGaneshReadAccess::CreateSkImage(
 
 sk_sp<SkImage>
 SkiaGaneshImageRepresentation::ScopedGaneshReadAccess::CreateSkImageForPlane(
-    int plane_index,
+    size_t plane_index,
     SharedContextState* context_state,
     SkImages::TextureReleaseProc texture_release_proc,
     SkImages::ReleaseContext release_context) {
@@ -755,7 +755,7 @@ SkiaGraphiteImageRepresentation::ScopedGraphiteReadAccess::CreateSkImage(
 }
 
 sk_sp<SkImage> SkiaGraphiteImageRepresentation::ScopedGraphiteReadAccess::
-    CreateSkImageForPlane(int plane_index,
+    CreateSkImageForPlane(size_t plane_index,
                           SharedContextState* context_state,
                           SkImages::TextureReleaseProc texture_release_proc,
                           SkImages::ReleaseContext release_context) {
