+++ gpu/command_buffer/service/shared_image/video_image_reader_image_backing.cc
@@ -256,8 +256,8 @@ class VideoImageReaderImageBacking::GLTextureVideoImageRepresentation
   GLTextureVideoImageRepresentation& operator=(
       const GLTextureVideoImageRepresentation&) = delete;
 
-  gles2::Texture* GetTexture(int plane_index) override {
-    DCHECK_EQ(plane_index, 0);
+  gles2::Texture* GetTexture(size_t plane_index) override {
+    DCHECK_EQ(plane_index, 0u);
     return texture_;
   }
 
@@ -303,8 +303,8 @@ class VideoImageReaderImageBacking::GLTexturePassthroughVideoImageRepresentation
       const GLTexturePassthroughVideoImageRepresentation&) = delete;
 
   const scoped_refptr<gles2::TexturePassthrough>& GetTexturePassthrough(
-      int plane_index) override {
-    DCHECK_EQ(plane_index, 0);
+      size_t plane_index) override {
+    DCHECK_EQ(plane_index, 0u);
     return passthrough_texture_;
   }
